fix(CentralSiteMigration): Use long queue for playbook
Big sites take time to migrate

+++ press/press/doctype/central_site_migration/central_site_migration.py
@@ -30,7 +30,7 @@ class CentralSiteMigration(Document):
 	def start(self):
 		self.status = "Pending"
 		self.save()
-		frappe.enqueue_doc(self.doctype, self.name, "_start", queue="short")
+		frappe.enqueue_doc(self.doctype, self.name, "_start", queue="long")
 
 	def _start(self):
 		try:
